pyproject.toml
[tool.poetry]
name = "mobile-tests-hw"
version = "0.1.0"
description = ""
authors = ["Maksim Kudaev <kudaev.m@nakukop.com>"]
readme = "README.md"
packages = [{include = "mobile_tests_hw"}]

[tool.poetry.dependencies]
python = "^3.11"
Appium-Python-Client = "^2.6.1"
selene = "^2.0.0b17"
allure-pytest = "^2.13.0"
pytest = "^7.2.2"
dynaconf = "^3.1.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = 1
log_cli_level = "INFO"
log_cli_format = "%(asctime)s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = " -vv -s --alluredir=allure-result --allure-no-capture "
